kubrick-api/pyproject.toml
[project]
name = "kubrick-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.0",
    "click>=8.2.1",
    "fastapi[standard]>=0.115.13",
    "fastmcp>=2.9.0",
    "groq>=0.28.0",
    "instructor>=1.9.0",
    "ipykernel>=6.29.5",
    "loguru>=0.7.3",
    "opencv-python>=4.12.0.88",
    "opik>=1.7.36",
    "pixeltable>=0.4.1",
    "pydantic-settings>=2.10.0",
    "ruff>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kubrick_api"]

[tool.ruff]
target-version = "py312"
